Union de topBar con CardGrid
