+++ src/components/EmpresaSidebar.tsx
@@ -7,7 +7,7 @@ interface EmpresaSidebarProps {
 
 const EmpresaSidebar: React.FC<EmpresaSidebarProps> = ({ empresas, onEmpresaSelect }) => {
     return (
-        <div style={{ width: '200px', backgroundColor: '#f1f1f1', padding: '10px' }}>
+        <div style={{ height:'100%', width: '200px', backgroundColor: '#f1f1f1', padding: '10px' }}>
         <button>Agregar Empresa</button>
         {empresas.map((empresa) => (
             <div key={empresa.id} style={{ marginTop: '10px' }}>
