--- src/components/screens/GridSucursales/CardGrid/CardGrid.module.css
@@ -1,35 +1,0 @@
-.containerSucPage{
-    display: grid;
-    justify-content: center;
-    align-items: center;
-    gap: 2vh;
-    height: 100%;
-    grid-template-columns: repeat(3,1fr);
-    padding: 1rem;
-    flex-grow: 1;
-
-}
-.containerSuc{
-    background-color: rgba(206, 225, 255, 0.53);
-    border-radius: 1rem;
-    border: 1px solid #121212;
-    box-shadow: 5px 4px 10px 2px black;
-
-}
-
-.containerTextSucPage{
-    padding: 0.5rem;
-    margin-top: 0.5rem;
-}
-
-.containerImgSucPage{
-    padding: 0.4rem;
-    height: 15rem;
-}
-
-.containerImgSucPage Img{
-    width: 10rem; /* Ajusta el ancho al 100% del contenedor */
-    height: auto; /* Mantiene la relación de aspecto */
-    border-radius: 0.4rem;
-}
-
