+++ src/components/screens/screen.css
@@ -1,0 +1,15 @@
+.container-screen{
+    margin: 1rem;
+    display: flex;
+    gap: 1rem;
+    height: 86vh;
+    overflow: hidden;
+}
+
+.featured-home{
+    padding: 2px;
+    height: 100%;
+    width: 100%;
+    border: solid 1px rgba(217, 217, 217, 1);
+    border-radius: 8px;
+}
