+++ src/components/styles/uiStyles/uiStyles.css
@@ -1,10 +1,10 @@
 .navBar{
     background-color: black;
-    height: 4rem;
+    height: 3.2rem;
 }
 
 .footer {
-    height: 4rem; 
+    height: 3.2rem; 
     background-color: black;
     position: fixed;
     width: 100%;
