+++ src/components/ui/featureds/featuredCardGrid/CardGrid.module.css
@@ -1,0 +1,50 @@
+.containerSucPage{
+    display: grid;
+    justify-content: center;
+    gap: 1.5rem;
+    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+    padding: 1rem;
+    flex-grow: 1;
+    height: 100%;
+    overflow-y: auto;
+    align-items: start; 
+}
+
+.containerSuc{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    min-width: 200px;
+    max-width: 350px;
+    background-color: rgba(206, 225, 255, 0.53);
+    border-radius: 8px;
+    border: 1px solid rgba(217, 217, 217, 1);
+    box-shadow: 4px 4px 4px rgba(196, 196, 196, 1);
+    text-align: center;
+    padding: 5px;
+    max-height: 450px;
+}
+
+.containerImgSucPage{
+    width: 90%;
+    height: 250px;
+    overflow: hidden;
+    border-radius: 5px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.containerImgSucPage Img{
+    max-width: 100%;
+    max-height: 100%;
+    object-fit: contain;
+}
+
+.containerTextSucPage{
+    margin-top: 10px;
+    font-size: 1rem;
+    line-height: 10px;
+}
+
