--- src/types/IEmpresa.ts
@@ -1,8 +1,0 @@
-import { ISucursal } from "../types/ISucursal"
-
-export interface IEmpresa {
-id: number;
-nombre: string;
-ciudad: string;
-sucursales: ISucursal[];
-}
