--- src/types/ISucursal.ts
@@ -1,6 +1,0 @@
-export interface ISucursal {
-    id: string;
-    nombre: string;
-    direccion: string;
-    apertura: string;
-}
